% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci_sequence.R
\name{fibonacci_sequence}
\alias{fibonacci_sequence}
\title{Fibonnaci sequence generator:}
\usage{
fibonacci_sequence(n)
}
\description{
param n>2 desired number of fibonacci numbers to be returned in a vector
   of length n
}
